Making methods to make Advertisements controller work

+++ src/main/java/ru/skypro/homework/repository/AdRepository.java
@@ -4,6 +4,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ru.skypro.homework.model.Ad;
 
+/**
+ * <h2>Repository for Ad entities (advertisements)</h2>
+ */
 @Repository
-public interface AdRepository extends JpaRepository<Ad,Integer> {
+public interface AdRepository extends JpaRepository<Ad, Long> {
 }
